done/checked pretty much everything on the fixable list. 'using static' in the library bits for the controllers is a problem though. doesnt want to compile with it, don't know how to fix it.

+++ c# folder/C# swingame/src/Model/Ship.cs
@@ -2,8 +2,8 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Data;
 using System.Diagnostics;
+
 /// <summary>
 /// A Ship has all the details about itself. For example the shipname,
 /// size, number of hits taken and the location. Its able to add tiles,
@@ -79,7 +79,7 @@ public class Ship
 		_tiles = new List<Tile>();
 
 		//gets the ship size from the enumarator
-		_sizeOfShip = _shipName;
+		_sizeOfShip = (int)_shipName;
 	}
 
 	/// <summary>
@@ -132,3 +132,4 @@ public class Ship
 		_direction = direction;
 	}
 }
+
